+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/controller/ContextBrokerNotificationController.java
@@ -1,12 +1,10 @@
 package es.in2.dome.blockchain.connector.integration.contextbroker.controller;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import es.in2.dome.blockchain.connector.integration.contextbroker.exception.HashLinkException;
 import es.in2.dome.blockchain.connector.integration.contextbroker.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+
 
 @Slf4j
 @RestController
@@ -16,9 +14,10 @@ public class ContextBrokerNotificationController {
 
     private final NotificationService notificationService;
 
+
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void receiveNotification(@RequestBody String notification) throws JsonProcessingException, HashLinkException {
+    public void receiveNotification(@RequestBody String notification) {
         log.debug("Notification received: {}", notification);
         notificationService.processNotification(notification);