+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/domain/dto/DomeEventDTO.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class BlockchainEventDTO {
+public class DomeEventDTO {
 
     private String type;
     private String data;