+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/domain/entity/DomeEventEntityMetadata.java
@@ -3,14 +3,13 @@ package es.in2.dome.blockchain.connector.integration.contextbroker.domain.entity
 
 import lombok.*;
 
-import java.util.ArrayList;
 
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class BlockchainEventEntityMetadata {
+public class DomeEventEntityMetadata {
 
-   private ArrayList<String> metadata = new ArrayList<>();
+   private String metadata;
 }