+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/exception/JsonReadingException.java
@@ -1,0 +1,8 @@
+package es.in2.dome.blockchain.connector.integration.contextbroker.exception;
+
+public class JsonReadingException extends RuntimeException {
+
+    public JsonReadingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}