+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/service/DomeEventService.java
@@ -3,6 +3,6 @@ package es.in2.dome.blockchain.connector.integration.contextbroker.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import es.in2.dome.blockchain.connector.integration.contextbroker.exception.HashLinkException;
 
-public interface DomeEventFactoryService {
-     String createBlockchainEvent(String type, String notificationdata) throws JsonProcessingException, HashLinkException;
+public interface DomeEventService {
+     String createDomeEvent (String type, String notificationdata) throws JsonProcessingException, HashLinkException;
 }