+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/service/impl/DomeEventServiceImpl.java
@@ -2,11 +2,11 @@ package es.in2.dome.blockchain.connector.integration.contextbroker.service.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import es.in2.dome.blockchain.connector.integration.contextbroker.domain.entity.BlockchainEventEntity;
-import es.in2.dome.blockchain.connector.integration.contextbroker.domain.entity.BlockchainEventEntityMetadata;
+import es.in2.dome.blockchain.connector.integration.contextbroker.domain.entity.DomeEventEntity;
 import es.in2.dome.blockchain.connector.integration.contextbroker.exception.HashLinkException;
-import es.in2.dome.blockchain.connector.integration.contextbroker.service.DomeEventFactoryService;
-import es.in2.dome.blockchain.connector.utils.HashlinkUtils;
+import es.in2.dome.blockchain.connector.integration.contextbroker.service.DomeEventService;
+import es.in2.dome.blockchain.connector.utils.ApplicationUtils;
+import es.in2.dome.blockchain.connector.utils.BlockchainConnectorUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -14,29 +14,29 @@ import org.springframework.stereotype.Service;
 import java.security.NoSuchAlgorithmException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class DomeEventFactoryServiceImpl implements DomeEventFactoryService {
+public class DomeEventServiceImpl implements DomeEventService {
 
-    private final HashlinkUtils hashlinkUtils;
+    private final ApplicationUtils applicationUtils;
 
     @Override
-    public String createBlockchainEvent(String type, String notificationdata) throws JsonProcessingException, HashLinkException {
+    public String createDomeEvent(String type, String notificationdata) throws JsonProcessingException, HashLinkException {
         String timestamp = generateTimestamp();
         String dataHashLink = createHashLink(notificationdata);
 
-        BlockchainEventEntity blockchainEventEntity = BlockchainEventEntity.builder()
+        DomeEventEntity domeEventEntity = DomeEventEntity.builder()
                 .type(type)
                 .dataLocation(dataHashLink)
                 .timestamp(timestamp)
-                .metadata(new BlockchainEventEntityMetadata())
+                .metadata(new ArrayList<>())
                 .build();
 
         // Create ObjectMapper to convert to JSON
-        ObjectMapper objectMapper = new ObjectMapper();
-        String blockchainEventJson = objectMapper.writeValueAsString(blockchainEventEntity);
+        String blockchainEventJson = new ObjectMapper().writeValueAsString(domeEventEntity);
 
 
         log.debug("Blockchain Event JSON: " + blockchainEventJson);
@@ -50,8 +50,8 @@ public class DomeEventFactoryServiceImpl implements DomeEventFactoryService {
 
     private String createHashLink(String resourceData) throws HashLinkException {
         try {
-            String resourceHash = hashlinkUtils.calculateSHA256Hash(resourceData);
-            return "hl:" + resourceHash;
+            String resourceHash = applicationUtils.calculateSHA256Hash(resourceData);
+            return BlockchainConnectorUtils.HASHLINK_PREFIX + resourceHash;
         } catch (NoSuchAlgorithmException e) {
             throw new HashLinkException("Error creating Hashlink");
         }