+++ src/main/java/es/in2/dome/blockchain/connector/integration/contextbroker/service/impl/NotificationServiceImpl.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import es.in2.dome.blockchain.connector.integration.contextbroker.exception.HashLinkException;
+import es.in2.dome.blockchain.connector.integration.contextbroker.exception.JsonReadingException;
 import es.in2.dome.blockchain.connector.integration.contextbroker.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,25 +17,39 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class NotificationServiceImpl implements NotificationService {
 
-    private final DomeEventFactoryServiceImpl domeEventFactoryService;
-
+    private final DomeEventServiceImpl domeEventService;
 
     @Override
-    public void processNotification(String data) throws JsonProcessingException, HashLinkException {
+    public void processNotification(String data) {
         ObjectMapper objectMapper = new ObjectMapper();
 
-        // Parse the data to extract the @type field
-        JsonNode root = objectMapper.readTree(data);
-        String type = root.get("data").get(0).get("type").asText();
-        log.debug("@type: " + type);
+        try {
+            JsonNode root = objectMapper.readTree(data);
+            String type = root.get("data").get(0).get("type").asText();
+            log.debug("@type: " + type);
 
-        // Remove the @type field from the notification
-        ((ObjectNode) root).remove("type");
+            // Remove the @type field from the notification
+            ((ObjectNode) root).remove("type");
 
-        // Convert the remaining data to a JSON string
-        String restOfData = objectMapper.writeValueAsString(root);
-        log.debug("Rest of data: " + restOfData);
+            // Convert the remaining data to a JSON string
+            String restOfData = objectMapper.writeValueAsString(root);
+            log.debug("Rest of data: " + restOfData);
 
-        domeEventFactoryService.createBlockchainEvent(type, restOfData);
+            createDomeEvent(type, restOfData);
+        } catch (JsonProcessingException e) {
+            throw new JsonReadingException("Error processing notification", e);
+        }
+    }
+
+    private void createDomeEvent(String type, String data) {
+        try {
+            domeEventService.createDomeEvent(type, data);
+        } catch (JsonProcessingException e) {
+            throw new JsonReadingException("Error creating DomeEvent: JSON processing issue", e);
+        } catch (HashLinkException e) {
+            throw new HashLinkException("Error while creating hashlink", e);
+        }
     }
 }
+
+