+++ src/main/java/es/in2/dome/blockchain/connector/utils/BlockchainConnectorUtils.java
@@ -1,0 +1,13 @@
+package es.in2.dome.blockchain.connector.utils;
+
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class BlockchainConnectorUtils {
+
+    private BlockchainConnectorUtils() {
+
+    }
+    public static final String HASHLINK_PREFIX = "hl:";
+}